Add CV functionality to Teleop Code
When the CV button on the wheel is pressed and the throttle and wheel
are zeroed, the AlignToHatch CV code will start running, allowing the
drivers to use CV in the teleop period. Furthermore, if the drivers want
to take control at any point, the CV code will stop running when the
drivers release the CV button or move the throttle or wheel.