Update elevator test code to use either wheel or buttons for control
Controlling the elevator using the wheel and the buttons at the same
time does not currently work. This commit allows the tester to switch
between the two control methods using the autonomous and tele-op modes.
Furthermore, the elevator speed is now slowed when going down to avoid a
fast drop.